+++ Anarchy/REST/Guild/Member/Models/PartialGuildMember.cs
@@ -1,11 +1,16 @@
 ﻿using Newtonsoft.Json;
 using System;
-using System.Collections.Generic;
 
 namespace Discord
 {
     public class PartialGuildMember : Controllable, IDisposable
     {
+        public PartialGuildMember()
+        {
+            OnClientUpdated += (sender, e) => User.SetClient(Client);
+        }
+
+
         [JsonProperty("guild_id")]
         internal ulong GuildId { get; set; }
 
@@ -19,47 +24,14 @@ namespace Discord
         }
 
 
-        [JsonProperty("nick")]
-        public string Nickname { get; internal set; }
-
-
-        [JsonProperty("roles")]
-        protected List<ulong> _roles;
-
-        public IReadOnlyList<MinimalRole> Roles
-        {
-            get
-            {
-                var roles = new List<MinimalRole>();
-
-                foreach (var role in _roles)
-                    roles.Add(new MinimalRole(GuildId, role).SetClient(Client));
-
-                return roles;
-            }
-        }
-
-
-        [JsonProperty("joined_at")]
-        public DateTime? JoinedAt { get; private set; }
-
-        [JsonProperty("premium_since")]
-        public DateTime? BoostingSince { get; private set; }
-
-
-        [JsonProperty("mute")]
-        public bool Muted { get; protected set; }
-
-
-        [JsonProperty("deaf")]
-        public bool Deafened { get; protected set; }
+        [JsonProperty("user")]
+        public DiscordUser User { get; internal set; }
 
 
         public new void Dispose()
         {
+            User = null;
             base.Dispose();
-            Nickname = null;
-            _roles = null;
         }
     }
 }